# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dispute Observatory API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Start Run",
        "operationId": "start_run_api_v1_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Runs",
        "operationId": "get_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/rerun": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Rerun Run",
        "operationId": "rerun_run_api_v1_runs__run_id__rerun_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run Events",
        "operationId": "get_run_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Seq"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ref"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events/stream": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Stream Run Events Endpoint",
        "operationId": "stream_run_events_endpoint_api_v1_runs__run_id__events_stream_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/decision": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run Decision",
        "operationId": "get_run_decision_api_v1_runs__run_id__decision_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queue/runs": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Start Queue Run",
        "operationId": "start_queue_run_api_v1_queue_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queue/runs/{run_id}": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get Queue Run",
        "operationId": "get_queue_run_api_v1_queue_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Actor": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ActorKind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "name"
        ],
        "title": "Actor"
      },
      "ActorKind": {
        "type": "string",
        "enum": [
          "graph_node",
          "agent",
          "subagent",
          "tool",
          "memory",
          "sandbox",
          "harness",
          "evaluator"
        ],
        "title": "ActorKind"
      },
...
      "DecisionResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "record": {
            "additionalProperties": true,
            "type": "object",
            "title": "Record"
          },
          "field_provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Field Provenance"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "case_id",
          "record",
          "field_provenance"
        ],
        "title": "DecisionResponse"
      },
      "EventEnvelope": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "span_id": {
            "type": "string",
            "title": "Span Id"
          },
          "parent_span_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Span Id"
          },
          "ts_wall": {
            "type": "string",
            "format": "date-time",
            "title": "Ts Wall"
          },
          "ts_virtual": {
            "type": "string",
            "format": "date-time",
            "title": "Ts Virtual"
          },
          "actor": {
            "$ref": "#/components/schemas/Actor"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Refs"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeSnapshot"
          },
          "usage": {
            "$ref": "#/components/schemas/EventUsage"
          },
          "redactions": {
            "items": {
              "$ref": "#/components/schemas/Redaction"
            },
            "type": "array",
            "title": "Redactions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "run_id",
          "case_id",
          "seq",
          "span_id",
          "parent_span_id",
          "ts_wall",
          "ts_virtual",
          "actor",
          "type",
          "summary",
          "payload",
          "runtime"
        ],
        "title": "EventEnvelope"
      },
      "EventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventEnvelope"
            },
            "type": "array",
            "title": "Items"
          },
          "next_after_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next After Seq"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_after_seq",
          "limit"
        ],
        "title": "EventPage"
      },
      "EventUsage": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "reasoning_tokens": {
            "type": "integer",
            "title": "Reasoning Tokens",
            "default": 0
          },
          "cached_tokens": {
            "type": "integer",
            "title": "Cached Tokens",
            "default": 0
          },
          "cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Usd",
            "default": "0"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "default": 0
          }
        },
        "type": "object",
        "title": "EventUsage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueueRunCreateRequest": {
        "properties": {
          "adapter": {
            "type": "string",
            "enum": [
              "fake",
              "openai"
            ],
            "title": "Adapter",
            "default": "fake"
          }
        },
        "type": "object",
        "title": "QueueRunCreateRequest"
      },
      "QueueRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "suspended",
              "decided",
              "cancelled",
              "failed",
              "ranked"
            ],
            "title": "Status"
          },
          "ranking": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "ranking"
        ],
        "title": "QueueRunResponse"
      },
      "Redaction": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "replacement": {
            "type": "string",
            "title": "Replacement",
            "default": "[REDACTED]"
          }
        },
        "type": "object",
        "required": [
          "path",
          "category"
        ],
        "title": "Redaction"
      },
...
      "RunCreateRequest": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "adapter": {
            "type": "string",
            "enum": [
              "fake",
              "openai"
            ],
            "title": "Adapter",
            "default": "fake"
          },
          "auto_resume": {
            "type": "boolean",
            "title": "Auto Resume",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "case_id"
        ],
        "title": "RunCreateRequest"
      },
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor",
          "limit"
        ],
        "title": "RunPage"
      },
...
      "RunStartResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "suspended",
              "decided",
              "cancelled",
              "failed",
              "ranked"
            ],
            "title": "Status"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "case_id",
          "status",
          "events_url",
          "stream_url"
        ],
        "title": "RunStartResponse"
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "suspended",
              "decided",
              "cancelled",
              "failed",
              "ranked"
            ],
            "title": "Status"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "first_seq": {
            "type": "integer",
            "title": "First Seq"
          },
          "last_seq": {
            "type": "integer",
            "title": "Last Seq"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "last_event_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event At"
          },
          "virtual_now": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Virtual Now"
          },
          "wait": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait"
          },
          "decision_available": {
            "type": "boolean",
            "title": "Decision Available"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "case_id",
          "status",
          "event_count",
          "first_seq",
          "last_seq",
          "started_at",
          "last_event_at",
          "virtual_now",
          "wait",
          "decision_available"
        ],
        "title": "RunSummary"
      },
      "RuntimeSnapshot": {
        "properties": {
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "agent_runtime": {
            "type": "string",
            "title": "Agent Runtime"
          },
          "model_gateway": {
            "type": "string",
            "title": "Model Gateway"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "adapter_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Adapter Versions"
          }
        },
        "type": "object",
        "required": [
          "config_hash",
          "agent_runtime",
          "model_gateway",
          "provider",
          "model"
        ],
        "title": "RuntimeSnapshot"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}